requirements.txt
streamlit == 1.37.1
streamlit-option-menu == 0.4.0
pyyaml == 6.0.2
pandas == 2.2.1
requests == 2.32.3
